Refactor header component and update app layout
Replaced the main content in app.html with the new <app-header> component and router outlet for better modularity. Refactored header.ts to HeaderComponent with greeting logic and OnInit lifecycle, and updated header.spec.ts to provide proper unit tests for HeaderComponent.
