+++ cv-angular/src/app/header/header.spec.ts
@@ -1,27 +1,24 @@
-import { Component, OnInit } from '@angular/core';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { HeaderComponent } from './header';
 
-@Component({
-  selector: 'app-header',
-  templateUrl: './header.component.html',
-  styleUrls: ['./header.component.css']
-})
-export class HeaderComponent implements OnInit {
-  greeting = { text: '', icon: '' };
+describe('HeaderComponent', () => {
+  let component: HeaderComponent;
+  let fixture: ComponentFixture<HeaderComponent>;
 
-  ngOnInit() {
-    this.greeting = this.getGreetingAndIcon();
-  }
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [ HeaderComponent ]
+    })
+    .compileComponents();
+  });
 
-  getGreetingAndIcon() {
-    const hour = new Date().getHours();
-    if (hour >= 6 && hour < 12) {
-      return { text: '¡Buenos días!', icon: '<i class=\"fas fa-coffee\"></i>' };
-    }
-    if (hour >= 12 && hour < 19) {
-      return { text: '¡Buenas tardes!', icon: '<i class=\"fas fa-sun\"></i>' };
-    }
-    return { text: '¡Buenas noches!', icon: '<i class=\"fas fa-moon\"></i>' };
-  }
-}import { ComponentFixture, TestBed } from '@angular/core/testing';
+  beforeEach(() => {
+    fixture = TestBed.createComponent(HeaderComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
 
-
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
