+++ cv-angular/src/app/header/header.ts
@@ -1,11 +1,25 @@
-import { Component } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 
 @Component({
   selector: 'app-header',
-  imports: [],
   templateUrl: './header.html',
-  styleUrl: './header.css'
+  styleUrls: ['./header.css']
 })
-export class Header {
+export class HeaderComponent implements OnInit {
+  greeting = { text: '', icon: '' };
 
+  ngOnInit() {
+    this.greeting = this.getGreetingAndIcon();
+  }
+
+  getGreetingAndIcon() {
+    const hour = new Date().getHours();
+    if (hour >= 6 && hour < 12) {
+      return { text: '¡Buenos días!', icon: '<i class=\"fas fa-coffee\"></i>' };
+    }
+    if (hour >= 12 && hour < 19) {
+      return { text: '¡Buenas tardes!', icon: '<i class=\"fas fa-sun\"></i>' };
+    }
+    return { text: '¡Buenas noches!', icon: '<i class=\"fas fa-moon\"></i>' };
+  }
 }
